requirements.txt
httpx>=0.27.0
pydantic>=2.0.0
rich>=13.0.0
python-dotenv>=1.0.0
rapidfuzz>=3.0.0
